requirements.txt
flask==3.1.1
mysql-connector-python==9.3.0
apscheduler==3.11.0
sshtunnel==0.4.0
